+++ pkg/query/query_test.go
@@ -50,7 +50,7 @@ func Benchmark_Query_Merge(b *testing.B) {
 				reg,
 			)
 			require.NoError(b, err)
-			colDB, err := col.DB("parca")
+			colDB, err := col.DB(context.Background(), "parca")
 			require.NoError(b, err)
 
 			schema, err := parcacol.Schema()
@@ -96,13 +96,16 @@ func Benchmark_Query_Merge(b *testing.B) {
 			api := NewColumnQueryAPI(
 				logger,
 				tracer,
-				m,
 				getShareServerConn(b),
-				query.NewEngine(
-					memory.NewGoAllocator(),
-					colDB.TableProvider(),
+				parcacol.NewQuerier(
+					tracer,
+					query.NewEngine(
+						memory.DefaultAllocator,
+						colDB.TableProvider(),
+					),
+					"stacktraces",
+					m,
 				),
-				"stacktraces",
 			)
 			b.ResetTimer()
 
@@ -144,7 +147,7 @@ func Benchmark_ProfileTypes(b *testing.B) {
 
 	require.NoError(b, col.ReplayWALs(ctx))
 
-	colDB, err := col.DB("parca")
+	colDB, err := col.DB(context.Background(), "parca")
 	require.NoError(b, err)
 
 	table, err := colDB.GetTable("stacktraces")
@@ -162,13 +165,16 @@ func Benchmark_ProfileTypes(b *testing.B) {
 	api := NewColumnQueryAPI(
 		logger,
 		tracer,
-		m,
 		getShareServerConn(b),
-		query.NewEngine(
-			memory.NewGoAllocator(),
-			colDB.TableProvider(),
+		parcacol.NewQuerier(
+			tracer,
+			query.NewEngine(
+				memory.DefaultAllocator,
+				colDB.TableProvider(),
+			),
+			"stacktraces",
+			m,
 		),
-		"stacktraces",
 	)
 	b.ResetTimer()