+++ ui/packages/shared/components/src/MetricsGraph/index.tsx
@@ -1,3 +1,16 @@
+// Copyright 2022 The Parca Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 import React, {useEffect, useRef, useState} from 'react';
 import * as d3 from 'd3';
 import MetricsSeries from '../MetricsSeries';
@@ -101,7 +114,7 @@ function generateGetBoundingClientRect(contextElement: Element, x = 0, y = 0) {
       left: domRect.x + x,
       right: domRect.x + x,
       bottom: domRect.y + y,
-    } as ClientRect);
+    } as DOMRect);
 }
 
 const virtualElement: VirtualElement = {
@@ -114,7 +127,7 @@ const virtualElement: VirtualElement = {
       left: 0,
       right: 0,
       bottom: 0,
-    } as ClientRect;
+    } as DOMRect;
   },
 };
 
@@ -153,7 +166,7 @@ export const MetricsTooltip = ({
   useEffect(() => {
     if (contextElement != null) {
       virtualElement.getBoundingClientRect = generateGetBoundingClientRect(contextElement, x, y);
-      update?.();
+      void update?.();
     }
   }, [x, y, contextElement, update]);
 
@@ -308,7 +321,7 @@ export const RawMetricsGraph = ({
   const yScale = d3
     .scaleLinear()
     // tslint:disable-next-line
-    .domain([minY, maxY])
+    .domain([minY, maxY] as Iterable<d3.NumberValue>)
     .range([height - margin, 0]);
 
   const color = d3.scaleOrdinal(d3.schemeCategory10);
@@ -330,7 +343,7 @@ export const RawMetricsGraph = ({
       const pointIndex = d3.minIndex(distances);
       const minDistance = distances[pointIndex];
       return {
-        pointIndex: pointIndex,
+        pointIndex,
         distance: minDistance,
       };
     });
@@ -460,7 +473,7 @@ export const RawMetricsGraph = ({
 
     return {
       labels: [],
-      seriesIndex: seriesIndex,
+      seriesIndex,
       timestamp: sample[0],
       value: sample[1],
       x: xScale(sample[0]),