+++ ui/packages/shared/profile/src/ProfileViewWithData.tsx
@@ -1,3 +1,16 @@
+// Copyright 2022 The Parca Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 import {QueryServiceClient, QueryRequest_ReportType} from '@parca/client';
 
 import {useQuery} from './useQuery';
@@ -25,7 +38,7 @@ export const ProfileViewWithData = ({
     response: flamegraphResponse,
     error: flamegraphError,
   } = useQuery(queryClient, profileSource, QueryRequest_ReportType.FLAMEGRAPH_UNSPECIFIED, {
-    skip: currentView != 'icicle' && currentView != 'both',
+    skip: currentView !== 'icicle' && currentView !== 'both',
   });
 
   const {
@@ -33,7 +46,7 @@ export const ProfileViewWithData = ({
     response: topTableResponse,
     error: topTableError,
   } = useQuery(queryClient, profileSource, QueryRequest_ReportType.TOP, {
-    skip: currentView != 'table' && currentView != 'both',
+    skip: currentView !== 'table' && currentView !== 'both',
   });
 
   const sampleUnit = profileSource.ProfileType().sampleUnit;