+++ ui/packages/shared/profile/src/components/ProfileShareButton/index.tsx
@@ -1,3 +1,16 @@
+// Copyright 2022 The Parca Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 import {useState} from 'react';
 import {Button, Modal, useGrpcMetadata} from '@parca/components';
 import {Icon} from '@iconify/react';
@@ -30,7 +43,7 @@ const ProfileShareModal = ({
   const metadata = useGrpcMetadata();
   const isFormDataValid = () => true;
 
-  const handleSubmit: () => void = async () => {
+  const handleSubmit: () => Promise<void> = async () => {
     try {
       setLoading(true);
       const {response} = await queryClient.shareProfile(
@@ -76,7 +89,7 @@ const ProfileShareModal = ({
               className="w-fit mt-4"
               onClick={e => {
                 e.preventDefault();
-                handleSubmit();
+                void handleSubmit();
               }}
               disabled={loading || !isFormDataValid()}
               type="submit"