+++ ui/packages/shared/profile/src/useDelayedLoader.ts
@@ -1,3 +1,16 @@
+// Copyright 2022 The Parca Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 import {useEffect, useState} from 'react';
 
 interface DelayedLoaderOptions {
@@ -5,10 +18,10 @@ interface DelayedLoaderOptions {
 }
 
 const useDelayedLoader = (isLoading = false, options?: DelayedLoaderOptions) => {
-  const {delay = 500} = options || {};
+  const {delay = 500} = options ?? {};
   const [isLoaderVisible, setIsLoaderVisible] = useState<boolean>(false);
   useEffect(() => {
-    let showLoaderTimeout;
+    let showLoaderTimeout: ReturnType<typeof setTimeout>;
     if (isLoading && !isLoaderVisible) {
       // if the request takes longer than half a second, show the loading icon
       showLoaderTimeout = setTimeout(() => {
@@ -18,7 +31,7 @@ const useDelayedLoader = (isLoading = false, options?: DelayedLoaderOptions) => 
       setIsLoaderVisible(false);
     }
     return () => clearTimeout(showLoaderTimeout);
-  }, [isLoading]);
+  }, [isLoading, isLoaderVisible, delay]);
 
   return isLoaderVisible;
 };