+++ ui/packages/shared/store/src/store.ts
@@ -1,9 +1,30 @@
+// Copyright 2022 The Parca Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 import {configureStore, combineReducers} from '@reduxjs/toolkit';
 import profileReducer from './slices/profileSlice';
 import uiReducer from './slices/uiSlice';
 import storage from 'redux-persist/lib/storage';
-import {persistReducer, FLUSH, REHYDRATE, PAUSE, PERSIST, PURGE, REGISTER} from 'redux-persist';
-import {persistStore} from 'redux-persist';
+import {
+  persistReducer,
+  FLUSH,
+  REHYDRATE,
+  PAUSE,
+  PERSIST,
+  PURGE,
+  REGISTER,
+  persistStore,
+} from 'redux-persist';
 
 const rootReducer = combineReducers({
   ui: uiReducer,
@@ -38,7 +59,7 @@ const store = configureStore({
 });
 
 const defaultExports = () => {
-  let persistor = persistStore(store);
+  const persistor = persistStore(store);
   return {store, persistor};
 };